[CODE] Organize edges
